% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bikeData.R
\name{bikeData}
\alias{bikeData}
\title{Function to pull data from the API}
\usage{
bikeData(url, layer, auth)
}
\arguments{
\item{url}{api url}

\item{layer}{list layer, eg data}

\item{auth}{if authorization is requireed}
}
\description{
Function to pull data from the API
}
