% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanFields.R
\name{cleanFields}
\alias{cleanFields}
\title{Function to clean up the data frames so that they can be combined}
\usage{
cleanFields(df)
}
\arguments{
\item{df}{dataframe to clean up}
}
\description{
Function to clean up the data frames so that they can be combined
}
