% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeAPI.R
\name{executeAPI}
\alias{executeAPI}
\title{Check API and then pull data}
\usage{
executeAPI(company, url, auth = NA)
}
\arguments{
\item{company}{dockless bike company, eg "spin"}

\item{url}{api url}

\item{auth}{if authorization is required, default = NA}
}
\description{
Check API and then pull data
}
