% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bikeData.R
\name{bikeData}
\alias{bikeData}
\title{Function to pull data from the API}
\usage{
bikeData(company, url, layer = "data", auth = NA)
}
\arguments{
\item{company}{dockless bike company, eg "spin"}

\item{url}{api url}

\item{layer}{list layer, default = "data"}

\item{auth}{if authorization is required, default = NA}
}
\description{
Function to pull data from the API
}
