% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDocklessdata.R
\name{getDocklessdata}
\alias{getDocklessdata}
\title{Pull, clean, and save dockless API data}
\usage{
getDocklessdata(output_filepath = NULL)
}
\arguments{
\item{output_filepath}{folder directory where you want the file to save to; if NULL, will just be stored in R}
}
\description{
Pull, clean, and save dockless API data
}
